Separate cmd_server to separate file
